test: move visual tests to components
