+++ .storybook/manager.ts
@@ -19,9 +19,9 @@ addons.setConfig({
             "https://lumada-design.github.io/uikit/master",
           )
         ) {
-          // Only hide "Tests" stories in the production build
+          // Hide "Test" stories in the production build
           // We need them in other cases for Chromatic and development
-          return !item.title?.includes("Tests/");
+          return !item.name.startsWith("Test");
         }
         return true;
       },
