+++ docs/guides/styling/UtilityClasses.stories.tsx
@@ -5,6 +5,7 @@ import {
   HvTypography,
 } from "@hitachivantara/uikit-react-core";
 import {
+  Abacus,
   Alert,
   DropRightXS,
   InProgress,
@@ -62,5 +63,47 @@ export const Animations: StoryObj = {
       </HvButton>
       <InProgress className="animate-spin" />
     </div>
+  ),
+};
+
+export const Test: StoryObj = {
+  parameters: {
+    chromatic: { disableSnapshot: false },
+    docs: { disable: true },
+  },
+  render: () => (
+    <>
+      <HvTypography variant="title4">Z-Index & Colors</HvTypography>
+      <section className="flex bg-default [&>div]:-mr-2">
+        <div className="w-12 h-12 bg-positive z-modal" />
+        <div className="w-11 h-11 bg-warning z-docked" />
+        <div className="w-10 h-10 bg-warning_120 z-8" />
+        <div className="w-9 h-9 bg-base_dark" />
+        <div className="w-8 h-8 bg-negative -z-1" />
+      </section>
+      <br />
+      <HvTypography variant="title4">Radius & Spacing</HvTypography>
+      <section className="flex gap-xs">
+        {[
+          "rounded-none m-1",
+          "rounded-base my-sm",
+          "rounded mr-md",
+          "rounded-2",
+          "rounded-circle",
+        ].map((classes) => (
+          <div key={classes} className="bg-secondary p-xs">
+            <div className={`w-8 h-8 bg-primary ${classes}`} />
+          </div>
+        ))}
+      </section>
+      <br />
+      <HvTypography variant="title4">Text & icons</HvTypography>
+      <section>
+        <div className="flex items-center text-warning">
+          <Abacus color="currentColor" />
+          <span>Text goes here</span>
+        </div>
+      </section>
+    </>
   ),
 };
