+++ packages/core/src/Badge/Badge.stories.tsx
@@ -129,3 +129,22 @@ export const Accessibility: StoryObj<HvBadgeProps> = {
     />
   ),
 };
+
+export const Test: StoryObj = {
+  parameters: {
+    chromatic: { disableSnapshot: false },
+    docs: { disable: true },
+  },
+  render: () => (
+    <div style={{ display: "flex", gap: 60, flexWrap: "wrap" }}>
+      <HvBadge count={10} icon={<Alert />} />
+      <HvBadge showCount count={8} icon={<Alert />} />
+      <HvBadge showCount count={88} icon={<Alert />} />
+      <HvBadge showCount count={888} icon={<Alert />} />
+      <HvBadge count={0} text="Events" textVariant="label" />
+      <HvBadge count={10} text="Events" />
+      <HvBadge showCount count={10} maxCount={5} text="Events" />
+      <HvBadge showCount count={8} text="Events" textVariant="title4" />
+    </div>
+  ),
+};
