+++ packages/core/src/ColorPicker/ColorPicker.stories.tsx
@@ -255,3 +255,38 @@ export const ControlledColorPicker: StoryObj<HvColorPickerProps> = {
     );
   },
 };
+
+export const Test: StoryObj = {
+  parameters: {
+    chromatic: { disableSnapshot: false },
+    docs: { disable: true },
+  },
+  render: () => (
+    <div style={{ display: "flex", gap: 5 }}>
+      <div style={{ width: 268 }}>
+        <HvColorPicker label="Color" defaultExpanded />
+      </div>
+      <div style={{ width: 268 }}>
+        <HvColorPicker
+          label="Color"
+          showSavedColors={false}
+          defaultValue="#C62828"
+          defaultExpanded
+        />
+      </div>
+      <div style={{ width: "134px" }}>
+        <HvColorPicker
+          label="Color"
+          showSavedColors={false}
+          showCustomColors={false}
+          defaultValue="#F6941E"
+          defaultExpanded
+        />
+      </div>
+      <div style={{ marginLeft: 130 }}>
+        <HvColorPicker label="Color" iconOnly />
+      </div>
+      <HvColorPicker label="Color" iconOnly defaultValue="#59941B" />
+    </div>
+  ),
+};
