+++ packages/core/src/InlineEditor/InlineEditor.stories.tsx
@@ -6,6 +6,7 @@ import {
   HvInlineEditor,
   HvInlineEditorProps,
   HvInput,
+  HvTypography,
   HvTypographyVariants,
 } from "@hitachivantara/uikit-react-core";
 
@@ -14,6 +15,18 @@ const meta: Meta<HvInlineEditorProps> = {
   component: HvInlineEditor,
 };
 export default meta;
+
+const variants: HvTypographyVariants[] = [
+  "display",
+  "title1",
+  "title2",
+  "title3",
+  "title4",
+  "body",
+  "label",
+  "caption1",
+  "caption2",
+];
 
 export const Main: StoryObj<typeof HvInlineEditor<typeof HvInput>> = {
   args: {
@@ -46,17 +59,6 @@ export const LargeVariants = () => {
   const [value, setValue] = useState(
     "Very very very long text that is likely to be truncated",
   );
-  const variants: HvTypographyVariants[] = [
-    "display",
-    "title1",
-    "title2",
-    "title3",
-    "title4",
-    "body",
-    "label",
-    "caption1",
-    "caption2",
-  ];
 
   return (
     <HvContainer>
@@ -74,4 +76,41 @@ export const LargeVariants = () => {
       </HvGrid>
     </HvContainer>
   );
+};
+
+export const Test: StoryObj = {
+  parameters: {
+    chromatic: { disableSnapshot: false },
+    docs: { disable: true },
+  },
+  render: () => {
+    return (
+      <HvContainer>
+        <HvTypography variant="title3">Enabled</HvTypography>
+        <br />
+        <HvInlineEditor />
+        <br />
+        <HvInlineEditor showIcon />
+        <br />
+        <HvTypography variant="title3">Disabled</HvTypography>
+        <br />
+        <HvInlineEditor disabled />
+        <br />
+        <HvInlineEditor disabled showIcon />
+        <br />
+        <HvTypography variant="title3">Typography Variants</HvTypography>
+        <br />
+        <HvGrid container>
+          {variants.map((variant) => (
+            <HvGrid item key={variant} xs={12} sm={6} style={{ minHeight: 64 }}>
+              <HvInlineEditor
+                variant={variant}
+                value="Very very very long text that is likely to be truncated"
+              />
+            </HvGrid>
+          ))}
+        </HvGrid>
+      </HvContainer>
+    );
+  },
 };
