+++ packages/core/src/Stack/Stack.stories.tsx
@@ -257,3 +257,56 @@ export const WithNavigation: StoryObj<HvStackProps> = {
     );
   },
 };
+
+export const Test: StoryObj = {
+  parameters: {
+    chromatic: { disableSnapshot: false },
+    docs: { disable: true },
+  },
+  render: () => (
+    <>
+      <div style={{ display: "flex", gap: 20 }}>
+        <HvStack>
+          <Box>1</Box>
+          <Box>2</Box>
+          <Box>3</Box>
+          <Box>4</Box>
+        </HvStack>
+        <HvStack spacing="md">
+          <Box>1</Box>
+          <Box>2</Box>
+          <Box>3</Box>
+          <Box>4</Box>
+        </HvStack>
+        <HvStack divider>
+          <Box>1</Box>
+          <Box>2</Box>
+          <Box>3</Box>
+          <Box>4</Box>
+        </HvStack>
+        <HvStack divider spacing="md">
+          <Box>1</Box>
+          <Box>2</Box>
+          <Box>3</Box>
+          <Box>4</Box>
+        </HvStack>
+      </div>
+      <br />
+      <br />
+      <HvStack divider direction="row">
+        <Box>1</Box>
+        <Box>2</Box>
+        <Box>3</Box>
+        <Box>4</Box>
+      </HvStack>
+      <br />
+      <br />
+      <HvStack direction="row">
+        <Box>1</Box>
+        <Box>2</Box>
+        <Box>3</Box>
+        <Box>4</Box>
+      </HvStack>
+    </>
+  ),
+};
