+++ packages/core/src/Table/stories/TableHooks/TableHooks.stories.tsx
@@ -1,5 +1,20 @@
+import { useMemo } from "react";
 import { StoryObj } from "@storybook/react";
 import { expect, userEvent, within } from "@storybook/test";
+import {
+  HvTable,
+  HvTableBody,
+  HvTableCell,
+  HvTableColumnConfig,
+  HvTableContainer,
+  HvTableHead,
+  HvTableHeader,
+  HvTableRow,
+  HvTableSection,
+  useHvHeaderGroups,
+  useHvTable,
+  useHvTableSticky,
+} from "@hitachivantara/uikit-react-core";
 
 import { AlternativeLayout } from "./AlternativeLayout";
 import AlternativeLayoutRaw from "./AlternativeLayout?raw";
@@ -160,4 +175,105 @@ export const UseHvHeaderGroupsStory: StoryObj = {
 export const UseHvRowStateStory: StoryObj = {
   parameters: { docs: { source: { code: UseHvRowStateRaw } } },
   render: () => <UseHvRowState />,
+};
+
+/** This was created to test grouped headers with sticky columns */
+export const TestHeaders: StoryObj = {
+  parameters: {
+    chromatic: { disableSnapshot: false },
+    docs: { disable: true },
+    a11y: {
+      config: {
+        rules: [
+          // the th cells without data are hidden to the a11y tree,
+          // but axe-core doesn't correctly understand that
+          { id: "th-has-data-cells", enabled: false },
+        ],
+      },
+    },
+  },
+  render: () => {
+    const data = useMemo(
+      () => [
+        { name: "Paul", email: "a@a.com", v1: "123", v2: "123", test: "123" },
+        { name: "Chris", email: "a@a.com", v1: "123", v2: "123", test: "123" },
+        { name: "Marta", email: "a@a.com", v1: "123", v2: "123", test: "123" },
+        { name: "Sarah", email: "a@a.com", v1: "123", v2: "123", test: "123" },
+      ],
+      [],
+    );
+
+    type Data = (typeof data)[number];
+
+    const columns = useMemo<HvTableColumnConfig<Data>[]>(
+      () => [
+        { accessor: "name", Header: "Name", sticky: "left" },
+        { accessor: "email", Header: "Email", sticky: "left" },
+        {
+          Header: "Group",
+          columns: [
+            { accessor: "v1", Header: "Var 1" },
+            { accessor: "v2", Header: "Var 2" },
+          ],
+        },
+        { accessor: "test", Header: "Test", sticky: "right" },
+      ],
+      [],
+    );
+
+    const { getTableProps, getTableBodyProps, prepareRow, headerGroups, rows } =
+      useHvTable<Data>(
+        {
+          columns,
+          data,
+        },
+        useHvHeaderGroups,
+        useHvTableSticky,
+      );
+
+    return (
+      <HvTableSection>
+        <HvTableContainer tabIndex={0}>
+          <HvTable {...getTableProps()}>
+            <HvTableHead>
+              {headerGroups.map((headerGroup) => (
+                <HvTableRow
+                  {...headerGroup.getHeaderGroupProps()}
+                  key={headerGroup.getHeaderGroupProps().key}
+                >
+                  {headerGroup.headers.map((col) => (
+                    <HvTableHeader
+                      {...col.getHeaderProps()}
+                      key={col.getHeaderProps().key}
+                    >
+                      {col.render("Header")}
+                    </HvTableHeader>
+                  ))}
+                </HvTableRow>
+              ))}
+            </HvTableHead>
+            <HvTableBody {...getTableBodyProps()}>
+              {rows.map((row) => {
+                prepareRow(row);
+                const { key, ...rowProps } = row.getRowProps();
+
+                return (
+                  <HvTableRow key={key} {...rowProps}>
+                    {row.cells.map((cell) => (
+                      <HvTableCell
+                        {...cell.getCellProps()}
+                        key={cell.getCellProps().key}
+                      >
+                        {cell.render("Cell")}
+                      </HvTableCell>
+                    ))}
+                  </HvTableRow>
+                );
+              })}
+            </HvTableBody>
+          </HvTable>
+        </HvTableContainer>
+      </HvTableSection>
+    );
+  },
 };
