+++ packages/core/src/Tag/stories/Tag.stories.tsx
@@ -1,8 +1,11 @@
+import { css } from "@emotion/css";
 import { Meta, StoryObj } from "@storybook/react";
+import { expect, userEvent, within } from "@storybook/test";
 import {
   HvOverflowTooltip,
   HvTag,
   HvTagProps,
+  theme,
 } from "@hitachivantara/uikit-react-core";
 
 import { Selectable as SelectableStory } from "./Selectable";
@@ -223,4 +226,50 @@ export const SelectableControlled: StoryObj<HvTagProps> = {
     },
   },
   render: () => <SelectableControlledStory />,
+};
+
+export const Test: StoryObj = {
+  parameters: {
+    chromatic: { disableSnapshot: false },
+    docs: { disable: true },
+  },
+  // For visual testing
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+    const button = canvas.getByRole("button", {
+      name: /asset 2/i,
+      pressed: false,
+    });
+    await userEvent.click(button);
+    await expect(
+      canvas.getByRole("button", { name: /asset 2/i, pressed: true }),
+    ).toBeInTheDocument();
+  },
+  render: () => (
+    <div style={{ display: "flex", gap: 10, flexWrap: "wrap" }}>
+      <HvTag label="Informational" />
+      <HvTag label="Informational" disabled />
+      <HvTag label="This is a very very very very very very very very long text for a tag" />
+      <HvTag color="positive_20" label="Success" />
+      <HvTag color="negative_20" label="Warning" />
+      <HvTag color="warning_20" label="Error" />
+      <HvTag
+        label="Success"
+        color="positive_20"
+        onDelete={() => {
+          alert("On Delete Action");
+        }}
+      />
+      <HvTag label="Feat" type="categorical" />
+      <HvTag label="Feat" type="categorical" disabled />
+      <HvTag label="Docs" type="categorical" color="cat2" />
+      <HvTag label="Asset 1" selectable color="cat1" />
+      <HvTag
+        label="Asset 2"
+        selectable
+        color="negative"
+        classes={{ root: css({ color: theme.colors.negative_20 }) }}
+      />
+    </div>
+  ),
 };
